Move ramp control to separate file.
